// backend/src/interfaces/measurements.interface.ts
export interface IMeasurement {
  _id: string;
  age: number;
  months: number;
  redHeight: number;
  yellowHeight: number;
  greenHeight: number;
  redWeight: number;
  yellowWeight: number;
  greenWeight: number;
  redWidth: number;
  yellowWidth: number;
  greenWidth: number;
  createdAt: string;
  updatedAt: string;
}